requirements.txt
python-telegram-bot==13.7
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.0.1
